Running to 20 viscous relaxation times fixes assymetry
So when running the simulation only to 10 tau (relaxation time,
tau = ~10x maxwell time) the displacement by 5 tau before
unloading had not reached equilibrium at 1000 m. So the initial
elastic displacement with unloading was different to loading at
the beginning. Running to 20 tau (i.e. loading for 10 tau and
unloading for 10 tau) means that the displacement gets to
1000 m ok and the displacement is elastic.

including the prestress advection to the elastic and viscoelastic
equations does make about a 10% difference with wavelength = D
= 3000 km which is about expected as this term should be about 10%
of the other terms according to Cathles pg 38. Reducing the wavelength
by 5x gives only a two percent difference c.f the analytical solution
which neglects the prestress advection term. (I've made some more notes in
my diary)

Highlights the funny situation for this viscoelastic problem
where solving the elastic equations and really prestress advection
only leading order when wavelength = 30000km (~ circumference
of Earth) yet the term is essential in the viscoelastic
limit because it provides the isostatic restoring force,
i.e. it represents the free surface deflection that
balances the load.