Add stephan fix for interiorbc eta plus domain width
Added more general definition of strong bc in timestepper
to allow for interior bc and removed my hack of setting the
homogenous bc to the strong bcs.

Separately realised that there is an instability to do with
the width of the domain, wavelength and timestep. If the timestep
is too long and the wavelength too short cf the domain width
then after a few relaxation times (where eta is decaying
approximately exponentially as expected) suddenly eta blows up
(along with vel/pressure) and forms a half wavelength cos which
oscillates between timesteps...

In the 2012 paper Stephan has defined the domain width as lambda
so only ever modelling one wavelength. With this default then convergence
works perfectly, including dt = tau, which previously was blowing up
because of this error...

Maybe the boundary conditions on the side walls are not consistent
and this somehow blows up for wide domains... not sure