Uniform refinement getting closer to aspect...
Reasonable agreement with aspect by eye for a mesh with 10 km
horizontal resolution and 25 km vertical resolution (cf target
resolution in upper 100 km of 6 km for aspect).

But there are still some obvious grid features in the vertical displacement
along the surface...

running on gadi on 208 cores (2 nodes with max memory of 1000Gb for normalsr)
seems to be ok. takes 4 minute per timestep which is about 5.5 hours
for 250 years of simulation with a timestep of 2.5 years.

now going to try a horizontally unstructured mesh to refine near
the ice disk and then extruded down...