Short simulation matches aspect very well
Initially I had the wrong surface ID which was causing a small
difference in maximum displcaments e.g. Fig 3b Weerdesteijn et al 2023.

Changing Firedrake versions to use new input ordering for vom
seems to have changed the solution quite a bit (~20% error on the maximum
surface displacement for the short simulation - it looks like the
vertical displacement is dropping too quickly). The solver residuals
are about 2 orders of magnitude larger for the new Firedrake
so maybe something has changed with the solvers?

now going to try the longer simulation