Add runscript that works with strong bcs
Change made before Hobart to scale the equations by a constant factor (~ Maxwell time)
gets strong bcs to work. This removes the small `kink' in the corner of the box that seemed
odd with weak bcs.

This simulation also has the rho1 term in and the delta rho for the free surface.