Get compressible test working
small changes to lame parameters in analytical solution and
using trace not div for deviatoric stress.

The change that made things converge though was not including
the compressible buoyuancy part... maybe this was an assumption
not included in cathles solution?

Still needs smaller timestep than zhong case. maybe needs coupling...