Add plotting script for convergence
dont multiply local error by dt for elastic step. for the zhong case
this brings convergence back to 1 for elastic and not 1.5 because
not artificially reducing the error as the timestep gets smaller
for the elastic case.

bit weird that coupled elastic incompressible has much lower error
but not converging in time. I thought that internal variable
shouldnt converge in time but compressible versions do and
not coupled incompressible also converges... bit confusing...