Put tanh around normalised ice thickness in exterior density
this was causing the inversion to not reach a thickness of 2
for the larger ice sheet! Effectively higher density ice where
thickness was greater than 1!

Also turns out with this change the rotational nullspace was really
important to get fast convergence of the optimisation solver...
My old inheritance scripts only started working once i put the rotational
nullspace in. and likewise this script stopped working so well without
the rotational nullspace